requirements.txt
aniso8601==8.0.0
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-restx==0.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dotenv==0.17.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
uWSGI==2.0.20
Werkzeug==1.0.1
